Migrate plotting onto cnplot; refactor plot/ and shared helpers
- Replace HATCHet's local plotting stack with the cnplot library
  (colormaps, integer-CN profiles/legends, genome axis, 1D/2D scatter).
- Reorganize plot/ by command: plot_cn, plot_cnp_panel, plot_cluster_bins
  (was plot_1d2d), plot_compute_cn (merges plot_pool/plot_scaling_2d/
  plot_cn_tree), plot_utils (shared fonts + genome-axis). Delete
  plot_cn_utils and plot_common.
- Drop all allele-specific (ascn) plotting and the --plot_ascn flag.
- Remove the derived CNP/PROPS columns; cn_<clone> is canonical. Drop
  prepare_seg_ucn; read_seg_ucn_file returns (df, clones).
- Move file readers + override_solution to io_utils.py; move
  compute_expected_baf_fcn into utils.py; section-organize utils.py.
- Panel: display-name clone labels, inline "i (xx.xx%)" props, --show_ploidy,
  --min_prop; restrict genome axis to data chromosomes.
- Rename --keep_gap to --show_gap; show_* flags use BooleanOptionalAction.
- cluster-bins: require --region_bed and thread it through plot_rdr_baf.
