+++ Snakefile
@@ -84,6 +84,7 @@ rule run_cluster_bins:
         b_mfile=os.path.join(bb_dir, "bb.Ballele.npz"),
         t_mfile=os.path.join(bb_dir, "bb.Tallele.npz"),
         genome_size=config["genome_size"],
+        region_bed=config["region_bed"],
     output:
         bbc_dir=directory(bbc_dir),
         bbc=os.path.join(bbc_dir, "bulk.bbc"),
@@ -101,6 +102,7 @@ rule run_cluster_bins:
             --bb_dir {input.bb_dir} \
             --bbc_dir {output.bbc_dir} \
             --genome_size {input.genome_size} \
+            --region_bed {input.region_bed} \
             {params.args} > {log} 2>&1
         """
