--- src/hatchet/plot/plot_1d2d.py
@@ -1,878 +1,0 @@
-import os
-import logging
-import contextlib
-
-import pandas as pd
-import numpy as np
-import seaborn as sns
-import matplotlib.pyplot as plt
-from collections import OrderedDict
-from matplotlib.lines import Line2D
-from matplotlib.collections import LineCollection
-from matplotlib.backends.backend_pdf import PdfPages
-
-
-from scipy.stats import norm, beta as beta_dist, gaussian_kde
-from scipy.signal import find_peaks
-from adjustText import adjust_text
-
-logging.getLogger("adjustText").setLevel(logging.WARNING)
-
-from hatchet.utils import *
-from hatchet.plot.plot_utils import *
-
-
-##################################################
-def get_transparency(
-    bbc: pd.DataFrame, by: str, one_tail=0.25, tail_alpha=0.2, nontail_alpha=1.0
-) -> pd.Series:
-    def get_alpha_one_row(row, per_grp_thres):
-        rd, baf = row["RD"], row["BAF"]
-        rd_l, rd_r, baf_l, baf_r = per_grp_thres[row[by]]
-        if rd <= rd_l or rd >= rd_r or baf <= baf_l or baf >= baf_r:
-            return tail_alpha
-        else:
-            return nontail_alpha
-
-    def get_left_threshold(arr):
-        return arr[min(max(0, int(len(arr) * one_tail)), len(arr) - 1)]
-
-    def get_right_threshold(arr):
-        return arr[min(max(0, int(len(arr) * (1 - one_tail))), len(arr) - 1)]
-
-    per_grp_thres = {}
-    for key in bbc[by].unique():
-        df = bbc[bbc[by] == key]
-        rd = sorted(df["RD"].tolist())
-        baf = sorted(df["BAF"].tolist())
-        per_grp_thres[key] = [
-            get_left_threshold(rd),
-            get_right_threshold(rd),
-            get_left_threshold(baf),
-            get_right_threshold(baf),
-        ]
-
-    alphas = bbc.apply(func=lambda r: get_alpha_one_row(r, per_grp_thres), axis=1)
-    alphas.name = "transparency"
-
-    return alphas
-
-
-def get_abs_positions_ignore_gap(
-    bin_info: pd.DataFrame,
-    regions: pd.DataFrame,
-    chrs: list,
-    chr_shift=int(10e6),
-):
-    bin_info["abs_pos"] = 0
-    bin_info["abs_start"] = 0
-    bin_info["abs_end"] = 0
-
-    ch_offset = chr_shift
-    ch_coords = []  # chromosome bounderies
-    seg_coords = []  # inter-segment gaps
-
-    regions_chs = regions.groupby(by="#CHR", sort=False)
-    bins_chs = bin_info.groupby(by="#CHR", sort=False, observed=True)
-    for ch in chrs:
-        ch_coords.append(ch_offset)
-        regions_ch = regions_chs.get_group(ch)
-        bins_ch = bins_chs.get_group(ch)
-        for si in range(len(regions_ch)):
-            wl_segment = regions_ch.iloc[si]
-            wl_start = wl_segment["START"]
-            wl_end = wl_segment["END"]
-            seg_end = ch_offset + (wl_end - wl_start)
-
-            bins_seg = bins_ch.loc[
-                (bins_ch["START"] >= wl_start) & (bins_ch["END"] <= wl_end), :
-            ]
-            if bins_seg.empty:
-                ch_offset = seg_end
-                continue
-
-            # global bin coords
-            bin_starts = (bins_seg["START"] - wl_start + ch_offset).to_numpy()
-            bin_ends = (bins_seg["END"] - wl_start + ch_offset).to_numpy()
-
-            bin_info.loc[bins_seg.index, "abs_start"] = bin_starts
-            bin_info.loc[bins_seg.index, "abs_end"] = bin_ends
-            bin_info.loc[bins_seg.index, "abs_pos"] = (bin_starts + bin_ends) // 2
-            # update global offsets
-            ch_offset = seg_end
-            if si < len(regions_ch) - 1:
-                seg_coords.append(ch_offset)  # centromere offset
-    ch_coords.append(ch_offset)  # genome end
-    axis_start = chr_shift
-    axis_end = ch_coords[-1] + chr_shift
-    return bin_info, ch_coords, axis_start, axis_end, seg_coords
-
-
-def get_abs_positions_keep_gap(
-    bin_info: pd.DataFrame,
-    chrom_sizes: dict,
-    chrs: list,
-    chr_shift=int(10e6),
-):
-    """
-    chr_bounds: global chromosome boundary
-    axis_end:
-    """
-    chr_offsets = OrderedDict()
-    for i, ch in enumerate(chrs):
-        if i == 0:
-            chr_offsets[ch] = chr_shift
-        else:
-            prev_ch = chrs[i - 1]
-            offset = chr_offsets[prev_ch] + chrom_sizes[prev_ch]
-            chr_offsets[ch] = offset
-    ch_coords = list(chr_offsets.values()) + [
-        chr_offsets[chrs[-1]] + chrom_sizes[chrs[-1]]
-    ]
-    axis_start = chr_shift
-    axis_end = ch_coords[-1] + chr_shift
-    bin_info["abs_pos"] = bin_info.apply(
-        func=lambda r: chr_offsets[r["#CHR"]] + (r.START + r.END) // 2, axis=1
-    ).to_numpy()
-    bin_info["abs_start"] = bin_info.apply(
-        func=lambda r: chr_offsets[r["#CHR"]] + r.START, axis=1
-    ).to_numpy()
-    bin_info["abs_end"] = bin_info.apply(
-        func=lambda r: chr_offsets[r["#CHR"]] + r.END, axis=1
-    ).to_numpy()
-    return bin_info, ch_coords, axis_start, axis_end, []
-
-
-##################################################
-def plot_1d(
-    ax: plt.Axes,
-    sample: str,
-    bin_info: pd.DataFrame,
-    vals: np.ndarray,
-    regions: pd.DataFrame,
-    chrom_sizes: dict,
-    exp_colname=None,
-    exp_groups=None,
-    val_type="BAF",
-    colors=None,
-    hue=None,
-    palette=None,
-    alphas=None,
-    ylim=(0, 1),
-    ylab="value",
-    title=None,
-    plot_chrname=True,
-    ignore_gap=True,
-    chr_shift=int(10e6),
-    markersize=2.0,
-    exp_linewidth=1.5,
-    bd_linewidth=1,
-    linecolor=(0, 0, 0, 1),
-    show_legend=True,
-    rasterized=True,
-):
-    """
-    plot chrom-level 1D BAF/RDR/FCN values.
-    assert bin_info is sorted by chromosome and positions
-    If ignore_gap=True, ignore region outside regions (like centromeres)
-        and represented as dashed line.
-    """
-    bin_info = bin_info.copy(deep=True)
-    chrs = bin_info["#CHR"].unique().tolist()
-    if ignore_gap:
-        bin_info, ch_coords, axis_start, axis_end, seg_coords = (
-            get_abs_positions_ignore_gap(bin_info, regions, chrs, chr_shift=chr_shift)
-        )
-    else:
-        bin_info, ch_coords, axis_start, axis_end, seg_coords = (
-            get_abs_positions_keep_gap(bin_info, chrom_sizes, chrs, chr_shift)
-        )
-    # Drop bins not matched to any region (abs_pos still 0 from init)
-    if ignore_gap:
-        matched = bin_info["abs_pos"] > 0
-        if not matched.all():
-            n_drop = (~matched).sum()
-            logging.warning(f"plot_1d: dropping {n_drop} bins not matching any region")
-            bin_info = bin_info[matched].reset_index(drop=True)
-            vals = np.asarray(vals)[matched.to_numpy()]
-            if colors is not None:
-                colors = np.asarray(colors)[matched.to_numpy()]
-            if hue is not None:
-                hue = np.asarray(hue)[matched.to_numpy()]
-            if exp_groups is not None:
-                exp_groups = np.asarray(exp_groups)[matched.to_numpy()]
-            if alphas is not None:
-                alphas = np.asarray(alphas)[matched.to_numpy()]
-    ##################################################
-    sns.scatterplot(
-        x=bin_info["abs_pos"],
-        y=vals,
-        ax=ax,
-        s=markersize,
-        color=colors,
-        hue=hue,
-        palette=palette,
-        legend=(hue is not None) and show_legend,
-        edgecolor="none",
-        linewidth=0,
-    )
-    if alphas is not None and ax.collections:
-        scatter = ax.collections[-1]
-        face = scatter.get_facecolors()
-        if len(face) == 1:
-            face = np.tile(face, (len(alphas), 1))
-        if len(face) == len(alphas):
-            face[:, 3] = np.asarray(alphas)
-            scatter.set_facecolors(face)
-            scatter.set_alpha(None)
-    if rasterized:
-        for coll in ax.collections:
-            coll.set_rasterized(True)
-    if val_type == "BAF":
-        ax.hlines(
-            y=0.5,
-            xmin=axis_start,
-            xmax=axis_end,
-            colors="grey",
-            linestyle=":",
-            linewidth=bd_linewidth,
-        )
-
-    if exp_colname is not None:
-        exp_vals = bin_info[exp_colname].to_numpy()
-        abs_starts = bin_info["abs_start"].to_numpy()
-        abs_ends = bin_info["abs_end"].to_numpy()
-        grp = exp_groups if exp_groups is not None else exp_vals
-
-        exp_lines = []
-        for ch in bin_info["#CHR"].unique():
-            idx = np.where(bin_info["#CHR"] == ch)[0]
-            i = 0
-            while i < len(idx):
-                j = i + 1
-                while j < len(idx) and grp[idx[j]] == grp[idx[i]]:
-                    j += 1
-                exp_lines.append(
-                    [
-                        (abs_starts[idx[i]], exp_vals[idx[i]]),
-                        (abs_ends[idx[j - 1]], exp_vals[idx[i]]),
-                    ]
-                )
-                i = j
-        ax.add_collection(
-            LineCollection(
-                exp_lines, linewidth=exp_linewidth, colors=[linecolor] * len(exp_lines)
-            )
-        )
-
-    ax.vlines(
-        ch_coords,
-        ymin=0,
-        ymax=1,
-        transform=ax.get_xaxis_transform(),
-        linewidth=bd_linewidth,
-        colors="k",
-    )
-
-    ax.set_ylim(ylim[0], ylim[1])
-    ax.set_ylabel(ylab)
-    ax.set_title(title)
-    xticks = [
-        ch_coords[i] + (ch_coords[i + 1] - ch_coords[i]) // 2
-        for i in range(len(ch_coords) - 1)
-    ]
-    plt.setp(ax, xlim=(0, axis_end), xticks=xticks, xlabel="")
-    if plot_chrname:
-        ax.set_xticklabels(chrs, rotation=60, fontsize=8)
-
-    return
-
-
-##################################################
-def plot_2d(
-    sample: str,
-    bin_info: pd.DataFrame,
-    xvals: np.ndarray,
-    yvals: np.ndarray,
-    exp_xvals=None,
-    exp_yvals=None,
-    exp_labels=None,
-    clone_props=None,
-    alphas=None,
-    hue=None,
-    palette=None,
-    xlab="BAF",
-    ylab="RDR",
-    xlim=None,
-    ylim=None,
-    title=None,
-    label_clone=False,
-    markersize=2.0,
-    markersize_centroid=10,
-    marker_bd_width=0.8,
-    dpi=300,
-    transparent=False,
-    out_file=None,
-    rasterized=True,
-    filtered_ids=None,
-    balanced_ids=None,
-    display_min_clone_prop=None,
-):
-    """2D RDR-vs-BAF joint plot with per-cluster KDE marginals.
-
-    Draws a scatter plot on the joint axes (colored by ``hue``) and
-    unfilled KDE curves on the marginal axes — one line per cluster,
-    colored to match the scatter palette.
-
-    Tumor clones whose proportion in ``clone_props`` is below
-    ``display_min_clone_prop`` are dropped from the per-state CN labels and the
-    clone-prop legend; ``None`` shows all clones.
-    """
-
-    def _tumor_visible(j):
-        return (
-            display_min_clone_prop is None
-            or clone_props is None
-            or clone_props[j] >= display_min_clone_prop
-        )
-
-    g0 = sns.JointGrid(x=xvals, y=yvals, hue=hue, palette=palette, xlim=xlim, ylim=ylim)
-    g0.refline(x=0.50)
-    g0.plot_joint(sns.scatterplot, s=markersize, legend=False, edgecolors="none")
-    g0.plot_marginals(sns.kdeplot, common_norm=False, linewidth=0.8, fill=False)
-    scatter = g0.ax_joint.collections[0]
-    scatter.set_rasterized(rasterized)
-    scatter.set_antialiased(False)
-
-    if alphas is not None:
-        scatter.set_alpha(None)
-        colors_ = scatter.get_facecolors()  # reused in 1D plot
-        colors_[:, 3] = np.asarray(alphas)
-        scatter.set_facecolors(colors_)
-    g0_colors = scatter.get_facecolors()
-
-    if exp_labels is not None:
-        filtered_ids = filtered_ids or set()
-        balanced_ids = balanced_ids or set()
-        texts = []
-        vis_x, vis_y = [], []
-        seen_labels = set()  # collapse centroids sharing the same visible-clone label
-        for ci, cid in enumerate(exp_labels):
-            if filtered_ids and cid in filtered_ids:
-                continue
-            center_text = cid
-            fontdict = {"fontsize": 10}
-            if label_clone:
-                states = [
-                    f"({cid[j][0]},{cid[j][1]})"
-                    for j in range(1, len(cid))
-                    if _tumor_visible(j)
-                ]
-                if not states:
-                    states = [f"({x[0]},{x[1]})" for x in cid][1:]
-                states = list(dict.fromkeys(states))  # dedupe within label, keep order
-                if len(states) == 1:
-                    fontdict["fontweight"] = "bold"
-                    center_text = states[0]
-                else:
-                    center_text = ",".join(states)
-                if center_text in seen_labels:
-                    continue
-                seen_labels.add(center_text)
-            else:
-                center_text = cid
-            t = g0.ax_joint.text(exp_xvals[ci], exp_yvals[ci], center_text, **fontdict)
-            texts.append(t)
-            vis_x.append(exp_xvals[ci])
-            vis_y.append(exp_yvals[ci])
-        with open(os.devnull, "w") as _devnull, contextlib.redirect_stdout(_devnull):
-            adjust_text(
-                texts,
-                x=np.asarray(vis_x),
-                y=np.asarray(vis_y),
-                ax=g0.ax_joint,
-                arrowprops=dict(arrowstyle="-", color="black", lw=0.5),
-            )
-        ex, ey = np.asarray(exp_xvals), np.asarray(exp_yvals)
-        is_vis = np.array(
-            [not (filtered_ids and cid in filtered_ids) for cid in exp_labels]
-        )
-        is_bal = np.array(
-            [bool(balanced_ids and cid in balanced_ids) for cid in exp_labels]
-        )
-        for mask, marker in [(is_vis & ~is_bal, "o"), (is_vis & is_bal, "s")]:
-            if mask.any():
-                g0.ax_joint.scatter(
-                    x=ex[mask],
-                    y=ey[mask],
-                    facecolors="none",
-                    edgecolors="black",
-                    s=markersize_centroid,
-                    linewidth=marker_bd_width,
-                    marker=marker,
-                )
-
-    if clone_props is not None:
-        custom_handles = []
-        for i, prop in enumerate(clone_props):
-            if i != 0 and not _tumor_visible(i):
-                continue
-            lab = f"Normal: {prop:.3f}" if i == 0 else f"Clone {i}: {prop:.3f}"
-            custom_handles.append(Line2D([0], [0], alpha=0, label=lab))
-        g0.ax_joint.legend(
-            handles=custom_handles,
-            title="",
-            loc="best",
-            fontsize="small",
-            fancybox=True,
-            framealpha=0.7,
-            handlelength=0,
-            handletextpad=0,
-        )
-
-    g0.set_axis_labels(xlabel=xlab, ylabel=ylab)
-    g0.figure.suptitle(title)
-    plt.tight_layout()
-    if out_file is not None:
-        g0.savefig(out_file, dpi=dpi, bbox_inches="tight", transparent=transparent)
-        plt.close(g0.figure)
-    return g0.figure, g0_colors
-
-
-##################################################
-def _is_multimodal(obs, min_count=30):
-    """Return True if KDE of obs has >1 prominent peak."""
-    if len(obs) < min_count:
-        return False
-    try:
-        kde = gaussian_kde(obs)
-    except (np.linalg.LinAlgError, ValueError):
-        return False
-    grid = np.linspace(obs.min(), obs.max(), 200)
-    kde_vals = kde(grid)
-    peaks, _ = find_peaks(kde_vals, prominence=0.1 * kde_vals.max())
-    return len(peaks) > 1
-
-
-def plot_clusters(
-    cluster_labels: np.ndarray,
-    X_rdrs: np.ndarray,
-    X_bafs: np.ndarray,
-    rdr_means: np.ndarray,
-    rdr_vars: np.ndarray,
-    baf_means: np.ndarray,
-    baf_taus: np.ndarray,
-    tumor_samples: list,
-    out_file: str = None,
-    cluster_ids: np.ndarray = None,
-    log_rdr: bool = False,
-    pdf=None,
-    palette=None,
-    bin_info=None,
-    lim_baf: tuple = None,
-    lim_rdr: tuple = None,
-):
-    """Plot per-cluster joint BAF-vs-RDR scatter with marginals and QQ plots.
-
-    One PDF page per cluster with joint-grid panels (scatter + contour,
-    marginal histograms) on top and BAF/RDR QQ plots below.
-
-    Args:
-        cluster_labels: (N,) 0-indexed cluster assignments.
-        X_rdrs:         (N, M) RDR values (original scale).
-        X_bafs:         (N, M) phased BAF values in [0, 1].
-        rdr_means:      (K, M) fitted RDR means (log-space if log_rdr).
-        rdr_vars:       (K, M) fitted RDR variances (log-space if log_rdr).
-        baf_means:      (K, M) fitted BAF means (one mixture component).
-        baf_taus:       (K, M) fitted BAF dispersion per cluster per sample.
-        tumor_samples:  list of sample names.
-        out_file:       output PDF path (used only when pdf is None).
-        log_rdr:        if True, plot log(RDR) and overlay Gaussian in log-space.
-        pdf:            open PdfPages to append pages to; if None, opens out_file.
-    """
-    import warnings
-    from matplotlib.gridspec import GridSpec, GridSpecFromSubplotSpec
-
-    K = rdr_means.shape[0]
-    M = rdr_means.shape[1]
-    N_total = len(cluster_labels)
-    if cluster_ids is None:
-        cluster_ids = np.arange(K)
-
-    _close_pdf = pdf is None
-    if _close_pdf:
-        pdf = PdfPages(out_file)
-    with warnings.catch_warnings():
-        warnings.filterwarnings(
-            "ignore", message="divide by zero", category=RuntimeWarning
-        )
-        for ki, k in enumerate(cluster_ids):
-            mask = cluster_labels == k
-            n_bins = int(mask.sum())
-            if n_bins == 0:
-                continue
-
-            is_multimodal = False
-            for m in range(M):
-                baf_obs_check = X_bafs[mask, m]
-                if log_rdr:
-                    rdr_obs_check = np.log(np.clip(X_rdrs[mask, m], 1e-6, None))
-                else:
-                    rdr_obs_check = X_rdrs[mask, m]
-                if _is_multimodal(baf_obs_check) or _is_multimodal(rdr_obs_check):
-                    is_multimodal = True
-                    break
-
-            title = f"Cluster {k}  (n={n_bins}/{N_total})"
-            title_color = "red" if is_multimodal else "black"
-
-            fig = plt.figure(figsize=(7 * M, 12))
-            fig.suptitle(title, fontsize=14, y=0.99, color=title_color)
-            outer = GridSpec(
-                2, M, figure=fig, height_ratios=[3, 2], wspace=0.35, hspace=0.35
-            )
-
-            for m in range(M):
-                baf_obs = X_bafs[mask, m]
-                if log_rdr:
-                    rdr_obs = np.log(np.clip(X_rdrs[mask, m], 1e-6, None))
-                    mu_k = float(rdr_means[ki, m])
-                    var_k = float(rdr_vars[ki, m])
-                    ylab = "log(RDR)"
-                else:
-                    rdr_obs = X_rdrs[mask, m]
-                    mu_k = float(rdr_means[ki, m])
-                    var_k = float(rdr_vars[ki, m])
-                    ylab = "RDR"
-                sigma_k = np.sqrt(max(var_k, 1e-12))
-
-                p_k = float(baf_means[ki, m])
-                tau_m = float(baf_taus[ki, m])
-                a_param = p_k * tau_m
-                b_param = (1.0 - p_k) * tau_m
-
-                inner = GridSpecFromSubplotSpec(
-                    2,
-                    2,
-                    subplot_spec=outer[0, m],
-                    height_ratios=[1, 4],
-                    width_ratios=[4, 1],
-                    hspace=0.05,
-                    wspace=0.05,
-                )
-                ax_main = fig.add_subplot(inner[1, 0])
-                ax_top = fig.add_subplot(inner[0, 0], sharex=ax_main)
-                ax_right = fig.add_subplot(inner[1, 1], sharey=ax_main)
-                fig.add_subplot(inner[0, 1]).axis("off")
-
-                cluster_color = palette[ki] if palette is not None else "0.3"
-                ax_main.scatter(
-                    baf_obs,
-                    rdr_obs,
-                    s=4,
-                    alpha=0.3,
-                    color=cluster_color,
-                    rasterized=True,
-                )
-                if a_param > 0 and b_param > 0:
-                    xg = np.linspace(max(0.001, p_k - 0.3), min(0.999, p_k + 0.3), 150)
-                    yg = np.linspace(mu_k - 4 * sigma_k, mu_k + 4 * sigma_k, 150)
-                    Xg, Yg = np.meshgrid(xg, yg)
-                    log_joint = np.log(
-                        np.maximum(beta_dist.pdf(Xg, a_param, b_param), 1e-300)
-                    ) + norm.logpdf(Yg, mu_k, sigma_k)
-                    ax_main.contour(
-                        Xg,
-                        Yg,
-                        np.exp(log_joint),
-                        levels=6,
-                        colors="red",
-                        linewidths=0.8,
-                        alpha=0.7,
-                    )
-                ax_main.set_xlabel("BAF", fontsize=9)
-                ax_main.set_ylabel(ylab, fontsize=9)
-                if lim_baf is not None:
-                    ax_main.set_xlim(lim_baf)
-                if lim_rdr is not None:
-                    ax_main.set_ylim(lim_rdr)
-
-                ax_top.hist(
-                    baf_obs, bins=80, density=True, alpha=0.6, color="steelblue"
-                )
-                if a_param > 0 and b_param > 0:
-                    x_baf = np.linspace(0.001, 0.999, 300)
-                    ax_top.plot(
-                        x_baf,
-                        beta_dist.pdf(x_baf, a_param, b_param),
-                        "r-",
-                        lw=1.2,
-                        label=f"p={p_k:.3f} tau={tau_m:.0f}",
-                    )
-                    ax_top.legend(fontsize=7, loc="upper right")
-                ax_top.tick_params(labelbottom=False)
-                ax_top.set_title(tumor_samples[m], fontsize=10)
-
-                ax_right.hist(
-                    rdr_obs,
-                    bins=80,
-                    density=True,
-                    alpha=0.6,
-                    color="salmon",
-                    orientation="horizontal",
-                )
-                y_rdr = np.linspace(mu_k - 4 * sigma_k, mu_k + 4 * sigma_k, 300)
-                ax_right.plot(
-                    norm.pdf(y_rdr, mu_k, sigma_k),
-                    y_rdr,
-                    "r-",
-                    lw=1.2,
-                    label=f"mu={mu_k:.3f}\nvar={var_k:.4f}",
-                )
-                ax_right.legend(fontsize=7, loc="upper right")
-                ax_right.tick_params(labelleft=False)
-
-                inner_qq = GridSpecFromSubplotSpec(
-                    1,
-                    2,
-                    subplot_spec=outer[1, m],
-                    wspace=0.35,
-                )
-                n_obs = int(mask.sum())
-                theoretical_q = np.linspace(1 / (n_obs + 1), n_obs / (n_obs + 1), n_obs)
-
-                ax_baf_qq = fig.add_subplot(inner_qq[0, 0])
-                if a_param > 0 and b_param > 0:
-                    baf_sorted = np.sort(baf_obs)
-                    baf_theo = beta_dist.ppf(theoretical_q, a_param, b_param)
-                    ax_baf_qq.scatter(
-                        baf_theo, baf_sorted, s=1, alpha=0.3, color="steelblue"
-                    )
-                    qq_lo = min(baf_theo.min(), baf_sorted.min())
-                    qq_hi = max(baf_theo.max(), baf_sorted.max())
-                    ax_baf_qq.plot([qq_lo, qq_hi], [qq_lo, qq_hi], "r-", lw=1)
-                    ss_res = np.sum((baf_sorted - baf_theo) ** 2)
-                    ss_tot = np.sum((baf_sorted - baf_sorted.mean()) ** 2)
-                    r2_baf = 1 - ss_res / ss_tot if ss_tot > 0 else float("nan")
-                    ax_baf_qq.text(
-                        0.05,
-                        0.95,
-                        f"$R^2$={r2_baf:.4f}",
-                        transform=ax_baf_qq.transAxes,
-                        fontsize=9,
-                        va="top",
-                        ha="left",
-                    )
-                ax_baf_qq.set_xlabel("Theoretical (Beta)", fontsize=9)
-                ax_baf_qq.set_ylabel("Observed", fontsize=9)
-                ax_baf_qq.set_title("BAF QQ", fontsize=10)
-
-                ax_rdr_qq = fig.add_subplot(inner_qq[0, 1])
-                rdr_sorted = np.sort(rdr_obs)
-                rdr_theo = norm.ppf(theoretical_q, mu_k, sigma_k)
-                ax_rdr_qq.scatter(rdr_theo, rdr_sorted, s=1, alpha=0.3, color="salmon")
-                qq_lo = min(rdr_theo.min(), rdr_sorted.min())
-                qq_hi = max(rdr_theo.max(), rdr_sorted.max())
-                ax_rdr_qq.plot([qq_lo, qq_hi], [qq_lo, qq_hi], "r-", lw=1)
-                ss_res = np.sum((rdr_sorted - rdr_theo) ** 2)
-                ss_tot = np.sum((rdr_sorted - rdr_sorted.mean()) ** 2)
-                r2_rdr = 1 - ss_res / ss_tot if ss_tot > 0 else float("nan")
-                ax_rdr_qq.text(
-                    0.05,
-                    0.95,
-                    f"$R^2$={r2_rdr:.4f}",
-                    transform=ax_rdr_qq.transAxes,
-                    fontsize=9,
-                    va="top",
-                    ha="left",
-                )
-                ax_rdr_qq.set_xlabel("Theoretical (Gaussian)", fontsize=9)
-                ax_rdr_qq.set_ylabel("Observed", fontsize=9)
-                ax_rdr_qq.set_title(f"{ylab} QQ", fontsize=10)
-
-            fig.subplots_adjust(top=0.95)
-            pdf.savefig(fig)
-            plt.close(fig)
-
-    if _close_pdf:
-        pdf.close()
-        logging.info(f"cluster diagnostics saved to {out_file}")
-
-
-##################################################
-def plot_rdr_baf(
-    samples: list,
-    bin_info: pd.DataFrame,
-    baf_mat: np.ndarray,
-    rdr_mat: np.ndarray,
-    genome_file: str,
-    cluster_labels=None,
-    expected_rdrs=None,
-    expected_bafs=None,
-    unique_labels=None,
-    label_clone=False,
-    xlab="BAF",
-    ylab="RDR",
-    out_dir=None,
-    out_prefix="",
-    dpi=300,
-    transparent=False,
-    row_width=20,
-    row_height=4,
-    maxlim_rdr=100,
-    rasterized=True,
-    rdr_means=None,
-    rdr_vars=None,
-    baf_taus=None,
-    log_rdr=False,
-    filtered_ids=None,
-    balanced_ids=None,
-):
-    """
-    Plot BAF-RDR scatter 1D and 2D.
-    baf_mat and rdr_mat have shape (bins, samples).
-    Saves all 1D plots to one PDF and all 2D plots to another PDF,
-    one page per sample in the order given by samples.
-    """
-    plt.rcParams["pdf.fonttype"] = 42
-    plt.rcParams["ps.fonttype"] = 42
-    plt.rcParams["svg.fonttype"] = "none"
-
-    bin_info = bin_info.copy(deep=True)
-
-    chrom_sizes = read_genome_sizes(genome_file)
-    palette = None
-    _lbl_to_idx = None
-    if cluster_labels is not None:
-        palette = set_palette(num_colors=len(np.unique(cluster_labels)))
-        if unique_labels is not None:
-            _lbl_to_idx = np.empty(int(cluster_labels.max()) + 1, dtype=int)
-            for i, lbl in enumerate(unique_labels):
-                _lbl_to_idx[lbl] = i
-        if filtered_ids and unique_labels is not None:
-            gray = (0.75, 0.75, 0.75)
-            for i, lbl in enumerate(unique_labels):
-                if lbl in filtered_ids:
-                    palette[i] = gray
-
-    out_name = out_prefix.rstrip("_") if out_prefix else "plot"
-    pdf = PdfPages(os.path.join(out_dir, f"{out_name}.pdf"))
-
-    global_lim_baf = (0, 1) if np.max(baf_mat) > 0.5 else (0, 0.55)
-    global_max_rdr = int(np.ceil(np.max(rdr_mat)))
-    global_lim_rdr = (0, min(max(2, global_max_rdr), maxlim_rdr))
-
-    for si, sample in enumerate(samples):
-        logging.info(f"plot {sample}")
-        bafs = baf_mat[:, si]
-        rdrs = rdr_mat[:, si]
-        exp_bafs = expected_bafs[:, si] if expected_bafs is not None else None
-        exp_rdrs = expected_rdrs[:, si] if expected_rdrs is not None else None
-
-        lim_baf = (0, 1) if np.max(bafs) > 0.5 else (0, 0.55)
-        max_rdr = int(np.ceil(np.max(rdrs)))
-        if max_rdr > maxlim_rdr:
-            num_exceeded = np.sum(rdrs >= maxlim_rdr)
-            logging.warning(
-                f"there are {num_exceeded} bins having RDR exceed maxlim_rdr={maxlim_rdr}"
-            )
-        lim_rdr = (0, min(max(2, max_rdr), maxlim_rdr))
-
-        # Page 1: 2D plot
-        fig_2d, g0_colors = plot_2d(
-            sample,
-            bin_info,
-            bafs,
-            rdrs,
-            exp_xvals=exp_bafs,
-            exp_yvals=exp_rdrs,
-            exp_labels=unique_labels,
-            alphas=None,
-            hue=cluster_labels,
-            palette=palette,
-            label_clone=label_clone,
-            xlab=xlab,
-            ylab=ylab,
-            xlim=lim_baf,
-            ylim=lim_rdr,
-            title=f"sample={sample}",
-            dpi=dpi,
-            transparent=transparent,
-            rasterized=rasterized,
-            filtered_ids=filtered_ids,
-            balanced_ids=balanced_ids,
-        )
-        pdf.savefig(fig_2d, dpi=dpi, bbox_inches="tight")
-        plt.close(fig_2d)
-
-        # Page 2: 1D plot
-        fig, axes = plt.subplots(nrows=2, ncols=1, figsize=(row_width, row_height))
-        for i, [ctype, ylim, vals, exp_vals] in enumerate(
-            [[ylab, lim_rdr, rdrs, exp_rdrs], [xlab, lim_baf, bafs, exp_bafs]]
-        ):
-            exp_colname = None
-            if cluster_labels is not None:
-                exp_colname = f"exp-{ctype}"
-                bin_info[exp_colname] = exp_vals[_lbl_to_idx[cluster_labels]]
-            plot_1d(
-                axes[i],
-                sample,
-                bin_info,
-                vals,
-                None,
-                chrom_sizes,
-                exp_colname=exp_colname,
-                exp_groups=cluster_labels,
-                val_type=ctype,
-                colors=g0_colors,
-                hue=cluster_labels if i == 0 else None,
-                palette=palette if i == 0 else None,
-                ylim=ylim,
-                ylab=ctype,
-                plot_chrname=True,
-                ignore_gap=False,
-                rasterized=rasterized,
-            )
-        fig.suptitle(f"sample={sample}")
-        axes[0].grid(True, axis="y")
-        axes[0].grid(False, axis="x")
-        axes[1].grid(True, axis="y")
-        axes[1].grid(False, axis="x")
-        if cluster_labels is not None:
-            ncol = max(1, int(np.ceil(len(unique_labels) / 10)))
-            sns.move_legend(
-                axes[0],
-                "center left",
-                bbox_to_anchor=(1.01, 0.5),
-                frameon=False,
-                title=None,
-                ncol=ncol,
-                fontsize=8,
-                markerscale=6.0,
-            )
-        fig.subplots_adjust(right=0.82)
-        fig.tight_layout(rect=[0, 0, 0.82, 1])
-        pdf.savefig(fig, dpi=dpi, bbox_inches="tight")
-        plt.close(fig)
-
-    # Cluster-level diagnostic pages
-    if rdr_means is not None:
-        plot_clusters(
-            cluster_labels,
-            rdr_mat,
-            baf_mat,
-            rdr_means,
-            rdr_vars,
-            expected_bafs,
-            baf_taus,
-            samples,
-            cluster_ids=unique_labels,
-            log_rdr=log_rdr,
-            pdf=pdf,
-            palette=palette,
-            bin_info=bin_info,
-            lim_baf=global_lim_baf,
-            lim_rdr=global_lim_rdr,
-        )
-    pdf.close()
-    return
