+++ src/hatchet/plot/plot_compute_cn.py
@@ -1,18 +1,264 @@
-"""CNT clone-tree + CNP profile rendering."""
+"""Plotting for the compute-cn command.
+
+Solution-pool CNP panels, the scaling-factor 2D diagnostic, and clone-tree
+rendering — everything the compute-cn pipeline draws.
+"""
 
 from __future__ import annotations
 
+import os
+import re
+import logging
+
 import numpy as np
+import pandas as pd
 import matplotlib.pyplot as plt
 from matplotlib.backends.backend_pdf import PdfPages
 from matplotlib.collections import LineCollection
 
-from hatchet.plot.plot_cn_utils import (
-    plot_ascn_legend,
-    plot_ascn_profile,
-    plot_cnv_legend,
-    plot_cnv_profile,
-)
+from cnplot import annotate_landmarks, plot_cnv_profile, plot_scatter_2d, set_palette
+from hatchet.utils import sort_df_chr
+from hatchet.plot.plot_utils import build_genome_axis, use_editable_fonts
+
+
+def _clones_from_cn(df):
+    """Ordered clone names ("normal", "clone1", ...) from a df's cn_ columns."""
+    n = len([c for c in df.columns if c.startswith("cn_")])
+    return ["normal"] + [f"clone{i}" for i in range(1, n)]
+
+
+def _format_pool_label(tag):
+    """Convert 'pool_p0.05_s1' to 'p=0.05,s=1'."""
+    m = re.match(r"pool_p([^_]+)_s(\d+)", tag)
+    if m:
+        return f"p={m.group(1)},s={m.group(2)}"
+    return tag
+
+
+def _fmt_prop(v):
+    """Round proportion to 2 decimals as percent; literal '0' if zero."""
+    pct = round(v * 100, 2)
+    return "0" if pct == 0 else f"{pct}%"
+
+
+def plot_pool_cnp(
+    pool_instances,
+    genome_size,
+    region_bed,
+    out_dir,
+    sel_df=None,
+    segs=None,
+    title=None,
+    width=20,
+    height=1,
+    dpi=150,
+    solve_mode=None,
+    sample_names=None,
+    out_name="pool.pdf",
+):
+    """Plot pool CNP panel into out_dir.
+
+    For cnt_cd: renders each solution's tree as a separate PDF.
+    For cd/ilp: plots a single multi-row CNV profile panel of Pareto solutions.
+
+    Args:
+        pool_instances: {sol_id: {"imf_obj", "reg_obj", "cA", "cB", "u", ...}}.
+        region_bed: Path to the whitelist region BED file.
+        out_dir: Output directory for pool plots.
+        sel_df: Selection DataFrame from model_select_elbow_from_regularization.
+        segs: {sol_id: seg_df} pre-computed segmentation DataFrames.
+        title: Optional figure title.
+        width: Figure width in inches.
+        height: Height in inches per profile row.
+        dpi: Output resolution.
+        solve_mode: "cd", "ilp", "cnt_cd", etc.
+        sample_names: Sample names for cnt_cd rendering.
+        genome_size: Chromosome-sizes path for the genome axis.
+    """
+    os.makedirs(out_dir, exist_ok=True)
+
+    use_editable_fonts()
+    # Restrict the axis to chromosomes present in the solutions' segments.
+    keep = set()
+    for sdf in (segs or {}).values():
+        keep.update(sdf["#CHR"].unique())
+    genome_axis = build_genome_axis(
+        region_bed, genome_size, keep_chroms=keep or None
+    )
+
+    if solve_mode == "cnt_cd":
+        for sol_id, sol in pool_instances.items():
+            tree = sol.get("tree")
+            if tree is None:
+                continue
+            seg_info = sort_df_chr(segs[sol_id].copy(), pos="START")
+            out_file = os.path.join(out_dir, f"{sol_id}.pdf")
+            render_cnt_tree(
+                tree,
+                seg_info,
+                genome_axis,
+                out_file,
+                u=sol.get("u"),
+                sample_names=sample_names,
+            )
+        return
+
+    selected_ids = set()
+    if sel_df is not None:
+        selected_ids = set(
+            sel_df.loc[sel_df["selected"] == "*", "instance_id"].tolist()
+        )
+
+    pareto_ids = []
+    if sel_df is not None:
+        pareto_ids = sel_df.loc[sel_df["is_pareto"], "instance_id"].tolist()
+    else:
+        pareto_ids = sorted(pool_instances.keys())
+
+    entries = []
+    for sol_id in pareto_ids:
+        sol = pool_instances[sol_id]
+        is_selected = sol_id in selected_ids
+        entries.append((sol_id, segs[sol_id], sol["imf_obj"], is_selected))
+    entries.sort(key=lambda x: x[2])
+
+    if not entries:
+        logging.warning(f"plot_pool_cnp: no solutions to plot, skipping {out_dir}")
+        return
+
+    nrows = len(entries)
+    first_seg_df = entries[0][1]
+    n_clones = len([c for c in first_seg_df.columns if c.startswith("cn_")])
+    n_samples = first_seg_df["SAMPLE"].nunique()
+    row_h = height * max(1, n_clones - 1) + 0.2 * max(0, n_samples - 1)
+    fig, axes = plt.subplots(
+        nrows=nrows + 1,
+        ncols=1,
+        figsize=(width, row_h * nrows),
+        gridspec_kw={"height_ratios": [row_h] * nrows + [2 * height]},
+    )
+    fig.subplots_adjust(hspace=0.6 + 0.1 * max(0, n_samples - 1))
+    main_axes = axes[:-1] if nrows > 1 else [axes[0]]
+    ax_leg = axes[-1]
+
+    for i, (label, seg_df, obj, is_selected) in enumerate(entries):
+        seg_info_all = sort_df_chr(seg_df.copy(), pos="START")
+        clones = _clones_from_cn(seg_info_all)
+        samples = seg_info_all["SAMPLE"].unique().tolist()
+        seg_info = seg_info_all.loc[
+            seg_info_all["SAMPLE"] == samples[0], :
+        ].reset_index(drop=True)
+
+        plot_cnv_profile(
+            main_axes[i],
+            seg_info,
+            genome_axis,
+            ax_leg=(ax_leg if i == nrows - 1 else None),
+            plot_chrname=True,
+            show_prop=False,
+        )
+
+        short_label = _format_pool_label(str(label))
+        if is_selected:
+            short_label += " *"
+        prop_lines = []
+        for sid in samples:
+            sp = seg_info_all.loc[seg_info_all["SAMPLE"] == sid, :].reset_index(
+                drop=True
+            )
+            cps = sp[[f"u_{c}" for c in clones]].iloc[0].tolist()
+            prop_lines.append(f"{sid}:" + "|".join(_fmt_prop(c) for c in cps))
+        ylabel = f"{short_label}\nimf {round(obj, 2)}\n" + "\n".join(prop_lines)
+        color = "red" if is_selected else "black"
+        main_axes[i].set_ylabel(
+            ylabel, rotation=0, ha="right", va="center", color=color
+        )
+
+    if title:
+        main_axes[0].set_title(title)
+    out_file = os.path.join(out_dir, out_name)
+    plt.savefig(out_file, dpi=dpi, bbox_inches="tight")
+    plt.close()
+    logging.info(f"pool CNP panel saved to {out_file}")
+
+
+def plot_scaling_2d(
+    samples: list,
+    bbcs: pd.DataFrame,
+    segs: pd.DataFrame,
+    scaling: dict,
+    out_file: str,
+    markersize: float = 3.0,
+    markersize_centroid: float = 14,
+    dpi: int = 300,
+    transparent: bool = False,
+    maxlim_rdr: int = 10,
+):
+    """2D RDR-vs-BAF scatter anchoring the scaling inference from get_scaling_factor.
+
+    One PDF page per sample and WGD mode (noWGD = diploid, WGD = tetraploid,
+    drawn only when a WGD scaling was inferred). Bins are colored per cluster via
+    cnplot's plot_scatter_2d; anchor clusters are circled and annotated with their
+    inferred (a, b) clonal states via annotate_landmarks.
+    """
+    use_editable_fonts()
+
+    clusters = sorted(bbcs["CLUSTER"].unique().tolist())
+    palette = set_palette(num_colors=len(clusters))
+    pal = {str(c): palette[i] for i, c in enumerate(clusters)}
+
+    panels = [("noWGD", scaling["diploid"])]
+    if scaling.get("tetraploid") is not None:
+        panels.append(("WGD", scaling["tetraploid"]))
+
+    cent = segs.set_index(["#ID", "SAMPLE"])[["RD", "BAF"]]
+
+    pdf = PdfPages(out_file)
+    for sample in samples:
+        sub = bbcs[bbcs["SAMPLE"] == sample]
+        obs = pd.DataFrame({"BAF": sub["BAF"].to_numpy(), "RD": sub["RD"].to_numpy()})
+        obs["CLUSTER"] = sub["CLUSTER"].astype(str).to_numpy()
+
+        lim_baf = (0, 1) if obs["BAF"].max() > 0.5 else (0, 0.55)
+        lim_rdr = (0, min(max(2, int(np.ceil(obs["RD"].max()))), maxlim_rdr))
+
+        for label, info in panels:
+            landmarks = []
+            for c, (a, b) in info["clonal"].items():
+                if (c, sample) not in cent.index:
+                    continue
+                landmarks.append(
+                    {
+                        "x": cent.loc[(c, sample), "BAF"],
+                        "y": cent.loc[(c, sample), "RD"],
+                        "label": f"({a},{b})",
+                        "clonal": True,
+                    }
+                )
+
+            p = (info.get("purities") or {}).get(sample)
+            ptxt = f"  purity={p:.3f}" if p is not None else ""
+            grid = plot_scatter_2d(
+                obs,
+                xcol="BAF",
+                ycol="RD",
+                hue="CLUSTER",
+                palette=pal,
+                xlim=lim_baf,
+                ylim=lim_rdr,
+                xlabel="BAF",
+                ylabel="RDR",
+                title=f"sample={sample}  {label}{ptxt}",
+                refline_x=0.5,
+                markersize=markersize,
+            )
+            annotate_landmarks(grid.ax_joint, landmarks, markersize=markersize_centroid)
+            pdf.savefig(
+                grid.figure, dpi=dpi, bbox_inches="tight", transparent=transparent
+            )
+            plt.close(grid.figure)
+    pdf.close()
+    logging.info(f"scaling 2D scatter saved to {out_file}")
 
 
 def _display_name(v, tree):
@@ -26,8 +272,8 @@ def _display_name(v, tree):
 def _inorder(node, ch):
     if node not in ch:
         return [node]
-    l, r = ch[node]
-    return _inorder(l, ch) + [node] + _inorder(r, ch)
+    lc, r = ch[node]
+    return _inorder(lc, ch) + [node] + _inorder(r, ch)
 
 
 def _compute_x_weighted(node, ch, ec, x_parent=0):
@@ -65,8 +311,8 @@ def _draw_sample_tree(
             return node in inactive
         if node not in tumor_children:
             return False
-        l, r = tumor_children[node]
-        return _is_inactive(l) and _is_inactive(r)
+        lc, r = tumor_children[node]
+        return _is_inactive(lc) and _is_inactive(r)
 
     tumor_root_sy = _sy(node_y[tumor_subtree_root])
     normal_y = tumor_y_offset + tumor_y_scale + 0.06
@@ -109,9 +355,9 @@ def _draw_sample_tree(
         zorder=10,
     )
 
-    for parent, (l, r) in tumor_children.items():
+    for parent, (lc, r) in tumor_children.items():
         px, py = x_weighted[parent], _sy(node_y[parent])
-        for c in (l, r):
+        for c in (lc, r):
             cx, cy = x_weighted[c], _sy(node_y[c])
             style = ":" if _is_inactive(c) else "-"
             color = "#BBBBBB" if _is_inactive(c) else "black"
@@ -169,13 +415,12 @@ def _draw_sample_tree(
 def render_cnt_tree(
     labeled_tree,
     bin_info,
-    regions,
+    genome_axis,
     out_path,
     u=None,
     sample_names=None,
     min_prop=0.03,
     fontsize=10,
-    plot_ascn=False,
 ):
     """Render a LabeledCloneTree as a multi-page PDF.
 
@@ -184,8 +429,8 @@ def render_cnt_tree(
 
     Args:
         labeled_tree: LabeledCloneTree with inferred CN and events.
-        bin_info: DataFrame with #CHR, START, END, CNP, PROPS.
-        regions: DataFrame with #CHR, START, END.
+        bin_info: DataFrame with #CHR, START, END, and cn_clone* columns.
+        genome_axis: cnplot GenomeAxis to draw the profile on.
         out_path: output PDF path.
         u: (n_leaves, P) usage matrix. If provided, page 2 is generated.
         sample_names: list of P sample names.
@@ -212,9 +457,10 @@ def render_cnt_tree(
     x_weighted = _compute_x_weighted(tumor_subtree_root, tumor_children, edge_cost, 0)
     max_leaf_x = max((x_weighted[v] for v in tree.tumor_leaves), default=1) or 1
 
-    # Rebuild CNP to include all tree nodes (leaves + internal) in reverse
-    # inorder, since plot_cnv_profile renders entries bottom-to-top.
+    # Reverse-inorder clone order maps each node to the same row plot_cnv_profile
+    # drew it in under the former CNP layout (entries stacked bottom-to-top).
     cnp_rev = list(reversed(cnp_nodes))
+    node_clones = [f"n{v}" for v in cnp_rev]
     leaf_cn_to_seg = {}
     for s in range(tree.a_all.shape[0]):
         key = tuple(
@@ -223,24 +469,18 @@ def render_cnt_tree(
         leaf_cn_to_seg[key] = s
     clone_cols = [f"cn_clone{i}" for i in range(1, len(tree.tumor_leaves) + 1)]
     bin_info = bin_info.copy()
-    new_cnps = []
+    seg_idx = []
     for _, row in bin_info.iterrows():
         key = tuple(
             (int(v.split("|")[0]), int(v.split("|")[1]))
             for v in (row[c] for c in clone_cols)
         )
-        s = leaf_cn_to_seg[key]
-        parts = [
-            f"{int(tree.a_all[s, tree.normal_leaf])}|{int(tree.b_all[s, tree.normal_leaf])}"
-        ]
-        for v in cnp_rev:
-            parts.append(f"{int(tree.a_all[s, v])}|{int(tree.b_all[s, v])}")
-        new_cnps.append(";".join(parts))
-    bin_info["CNP"] = new_cnps
-    prop_parts = [str(np.mean(tree.node_props[tree.normal_leaf]))]
-    for v in cnp_rev:
-        prop_parts.append(str(np.mean(tree.node_props[v])))
-    bin_info["PROPS"] = ";".join(prop_parts)
+        seg_idx.append(leaf_cn_to_seg[key])
+    seg_idx = np.asarray(seg_idx)
+    for v in cnp_nodes:
+        a = tree.a_all[seg_idx, v].astype(int)
+        b = tree.b_all[seg_idx, v].astype(int)
+        bin_info[f"cn_n{v}"] = [f"{ai}|{bi}" for ai, bi in zip(a, b)]
 
     with PdfPages(out_path) as pdf:
         # Page 1
@@ -257,15 +497,13 @@ def render_cnt_tree(
         ax_cn = fig.add_subplot(gs[0, 1])
         ax_leg = fig.add_subplot(gs[1, :])
 
-        _profile_fn = plot_ascn_profile if plot_ascn else plot_cnv_profile
-        _profile_fn(
+        plot_cnv_profile(
             ax_cn,
             bin_info,
-            regions,
-            width=20,
-            height=1,
+            genome_axis,
+            ax_leg=ax_leg,
+            clones=node_clones,
             show_prop=False,
-            show_clone_name=False,
         )
         ax_cn.set_yticks([(num_rows - 1 - i + 0.5) * h for i in range(num_rows)])
         ax_cn.set_yticklabels(
@@ -284,9 +522,9 @@ def render_cnt_tree(
         ax_tree.set_xlim(-0.5, max_leaf_x + 2.0)
         ax_tree.set_ylim(0, 1)
         ax_tree.axis("off")
-        for parent, (l, r) in tumor_children.items():
+        for parent, (lc, r) in tumor_children.items():
             px, py = x_weighted[parent], node_y[parent]
-            for c in (l, r):
+            for c in (lc, r):
                 cx, cy = x_weighted[c], node_y[c]
                 ax_tree.plot([px, px], [py, cy], "k-", lw=2)
                 ax_tree.plot([px, cx], [cy, cy], "k-", lw=2)
@@ -318,8 +556,6 @@ def render_cnt_tree(
                     zorder=10,
                 )
 
-        _legend_fn = plot_ascn_legend if plot_ascn else plot_cnv_legend
-        _legend_fn(ax_leg)
         pdf.savefig(fig, bbox_inches="tight", dpi=150)
         plt.close(fig)
